Move grpc helper to utils
