+++ go/arrow/flight/flightsql/driver/driver.go
@@ -19,7 +19,6 @@ import (
 	"context"
 	"database/sql"
 	"database/sql/driver"
-	"encoding/base64"
 	"errors"
 	"fmt"
 	"io"
@@ -488,37 +487,6 @@ func (c *Connection) BeginTx(ctx context.Context, opts sql.TxOptions) (driver.Tx
 	c.txn = tx
 
 	return &Tx{tx: tx, timeout: c.timeout}, nil
-}
-
-// *** Internal functions and structures ***
-type grpcCredentials struct {
-	username string
-	password string
-	token    string
-	params   map[string]string
-}
-
-func (g grpcCredentials) GetRequestMetadata(ctx context.Context, uri ...string) (map[string]string, error) {
-	md := make(map[string]string, len(g.params)+1)
-
-	// Authentication parameters
-	switch {
-	case g.token != "":
-		md["authorization"] = "Bearer " + g.token
-	case g.username != "":
-
-		md["authorization"] = "Basic " + base64.StdEncoding.EncodeToString([]byte(g.username+":"+g.password))
-	}
-
-	for k, v := range g.params {
-		md[k] = v
-	}
-
-	return md, nil
-}
-
-func (g grpcCredentials) RequireTransportSecurity() bool {
-	return g.token != "" || g.username != ""
 }
 
 // Register the driver on load.
